+++ src/jtl/Connector/Shopware/Controller/CustomerOrder.php
@@ -62,13 +62,14 @@ class CustomerOrder extends DataController
 
                     // Billsafe
                     if ($paymentModuleCode === PaymentTypes::TYPE_BILLSAFE
-                    && isset($orderSW['attribute']['swagBillsafeIban'])
-                    && isset($orderSW['attribute']['swagBillsafeBic'])) {
+                        && isset($orderSW['attribute']['swagBillsafeIban'])
+                        && isset($orderSW['attribute']['swagBillsafeBic'])) {
                         $order->setPui(sprintf(
-                            'Bitte bezahlen Sie %s %s an folgendes Konto: %s',
+                            'Bitte bezahlen Sie %s %s an folgendes Konto: %s Verwendungszweck: BTN %s',
                             $orderSW['invoiceAmount'],
                             $order->getCurrencyIso(),
-                            sprintf('IBAN: %s, BIC: %s', $orderSW['attribute']['swagBillsafeIban'], $orderSW['attribute']['swagBillsafeBic'])
+                            sprintf('IBAN: %s, BIC: %s', $orderSW['attribute']['swagBillsafeIban'], $orderSW['attribute']['swagBillsafeBic']),
+                            $orderSW['transactionId']
                         ));
                     }